Labeling error
Phases space plot plot for the pendulum has the "Position" and "Velocity" labels are interchanged. Physically, it makes more sense for a pendulum to transition from "circular" to "sinusoidal" at higher velocity values as well. Proposed replacing with "Angular Velocity" and "Angular Position" for clarity, though "Velocity" and "Position" respectively probably suffices. Corresponding change is proposed to the graph in the docs website as well.